feat: trading-bot 클래스 변경, PositionManager 클래스 추가, 기타 리팩토링(미완성)
