+++ requirements.txt
@@ -1,3 +1,9 @@
 certifi==2024.12.14
 pip==24.3.1
 wheel==0.45.1
+
+Faker~=33.3.1
+pytest~=8.3.4
+pyupbit~=0.2.34
+PyMySQL~=1.1.1
+python-dotenv~=1.0.1
